backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
python-multipart
fastapi-clerk-auth
cryptography
psycopg[binary]
